Criando método alternaVisoes
